* Boxes are now drawn, with textures and colours (to eventually denote masses) * Added textures and config file (from which many options are parsed)

+++ 2DPhysicsDemo/units.h
@@ -5,5 +5,5 @@
 // This file defines the units the application will use. Using these variables, we can maintain consistency
 // in object sizes. For example, 3 meters is equal to 3*METER = meters(3)
 
-const f64 METER = 0.1;
-inline f64 meters(double in) { return in * METER; };
+const f32 METER = 0.1f;
+inline f32 meters(f32 in) { return in * METER; };
